requirements.txt
streamlit
openai
paddleocr
pillow  
altair  
pandas
python-multipart
paddlepaddle
opencv-python
python-dotenv 
google-genai
streamlit-aggrid
